% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZdogAnchor.R
\name{as.character.ZdogAnchor}
\alias{as.character.ZdogAnchor}
\title{Retrieve character representation of ZdogAnchor}
\usage{
\method{as.character}{ZdogAnchor}(x, ...)
}
\arguments{
\item{x}{ZdogAnchor object}

\item{...}{other arguments}
}
\description{
Retrieve character representation of ZdogAnchor
}
