% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZdogIllustration.R
\name{as.character.ZdogIllustration}
\alias{as.character.ZdogIllustration}
\title{Convert a ConvasDocument to a string}
\usage{
\method{as.character}{ZdogIllustration}(x, ...)
}
\arguments{
\item{x}{ZdogIllustration object}

\item{...}{other arguments}
}
\description{
Convert a ConvasDocument to a string
}
