% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZdogShapes.R
\docType{data}
\name{shape_specs}
\alias{shape_specs}
\title{Specification for all the shapes in Zdog}
\format{An object of class \code{list} of length 9.}
\usage{
shape_specs
}
\description{
This will be used to
\itemize{
  \item{build R6 classes}
  \item{add methods to the ZdogAnchor class}
  \item{create a ztag helper for creating shapes}
}
}
\keyword{datasets}
