% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZdogIllustration.R
\name{zdog_ill}
\alias{zdog_ill}
\title{Helper to create ZdogIllustration objects}
\usage{
zdog_ill(..., width = 240, height = 240)
}
\arguments{
\item{...}{Initialise the zdog doc with these elements}

\item{width, height}{Zdog dimensions}
}
\description{
Helper to create ZdogIllustration objects
}
