% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZdogAnchor.R
\docType{data}
\name{ZdogAnchor}
\alias{ZdogAnchor}
\alias{zdog_anchor}
\title{Zdog element builder}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ZdogAnchor
}
\description{
Zdog element builder
}
\keyword{datasets}
