% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZdogIllustration.R
\docType{data}
\name{ZdogIllustration}
\alias{ZdogIllustration}
\title{HTML5 Zdog Builder}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ZdogIllustration
}
\description{
HTML5 Zdog Builder
}
\keyword{datasets}
