% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZdogShape.R, R/ZdogShapes.R, R/ZdogText.R
\docType{data}
\name{ZdogShapeClass}
\alias{ZdogShapeClass}
\alias{ZdogRect}
\alias{ZdogRoundedRect}
\alias{ZdogEllipse}
\alias{ZdogPolygon}
\alias{ZdogShape}
\alias{ZdogHemisphere}
\alias{ZdogCone}
\alias{ZdogCylinder}
\alias{ZdogBox}
\alias{ZdogText}
\title{Zdog element builder}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ZdogShapeClass

ZdogText
}
\description{
Zdog element builder
}
\keyword{datasets}
