% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZdogShapes.R
\docType{data}
\name{ztag}
\alias{ztag}
\title{a helper for initializing Zdog classes}
\format{An object of class \code{list} of length 9.}
\usage{
ztag
}
\description{
a helper for initializing Zdog classes
}
\keyword{datasets}
